+++ aleph-client/src/staking.rs
@@ -1,9 +1,13 @@
-use crate::{send_xt, wait_for_session, BlockNumber, Connection};
+use codec::Compact;
 use log::info;
 use pallet_staking::{RewardDestination, ValidatorPrefs};
 use sp_core::Pair;
 use sp_runtime::Perbill;
-use substrate_api_client::{compose_call, compose_extrinsic, AccountId, GenericAddress, XtStatus};
+use substrate_api_client::{
+    compose_call, compose_extrinsic, AccountId, Balance, GenericAddress, XtStatus,
+};
+
+use crate::{get_locked_balance, send_xt, wait_for_session, BlockNumber, Connection, KeyPair};
 
 pub fn bond(
     connection: &Connection,
@@ -21,10 +25,10 @@ pub fn bond(
     send_xt(&connection, xt.hex_encode(), "bond", status);
 }
 
-pub fn validate(connection: &Connection, validator_commision_percentage: u8, status: XtStatus) {
+pub fn validate(connection: &Connection, validator_commission_percentage: u8, status: XtStatus) {
     let prefs = ValidatorPrefs {
         blocked: false,
-        commission: Perbill::from_percent(validator_commision_percentage as u32),
+        commission: Perbill::from_percent(validator_commission_percentage as u32),
     };
     let xt = compose_extrinsic!(connection, "Staking", "validate", prefs);
     send_xt(&connection, xt.hex_encode(), "validate", status);
@@ -91,4 +95,147 @@ fn wait_for_era_completion(
     let first_session_in_next_era = next_era_index * sessions_per_era;
     wait_for_session(connection, first_session_in_next_era)?;
     Ok(next_era_index)
+}
+
+pub fn payout_stakers(
+    stash_connection: &Connection,
+    stash_account: &AccountId,
+    era_number: BlockNumber,
+) {
+    let xt = compose_extrinsic!(
+        stash_connection,
+        "Staking",
+        "payout_stakers",
+        stash_account,
+        era_number
+    );
+
+    send_xt(
+        &stash_connection,
+        xt.hex_encode(),
+        "payout stakers",
+        XtStatus::InBlock,
+    );
+}
+
+pub fn check_non_zero_payouts_for_era(
+    stash_connection: &Connection,
+    accounts_to_check_balance: &[AccountId],
+    stash_account: &AccountId,
+    era: BlockNumber,
+) {
+    let locked_stash_balance_before_payout = accounts_to_check_balance
+        .iter()
+        .map(|account| get_locked_balance(account, stash_connection))
+        .collect::<Vec<_>>();
+    payout_stakers(stash_connection, stash_account, era - 1);
+    let locked_stash_balance_after_payout = accounts_to_check_balance
+        .iter()
+        .map(|account| get_locked_balance(account, stash_connection))
+        .collect::<Vec<_>>();
+    locked_stash_balance_before_payout.into_iter().zip(locked_stash_balance_after_payout.into_iter()).zip(accounts_to_check_balance.iter()).
+        for_each(|((balance_before, balance_after), account_id)| {
+            assert!(balance_after[0].amount > balance_before[0].amount,
+                    "Expected payout to be positive in locked balance for account {}. Balance before: {}, balance after: {}",
+                    account_id, balance_before[0].amount, balance_after[0].amount);
+        }
+        );
+}
+
+pub fn batch_bond(
+    connection: &Connection,
+    stash_controller_accounts: &[(&AccountId, &AccountId)],
+    bond_value: u128,
+    reward_destination: RewardDestination<GenericAddress>,
+) {
+    let batch_bond_calls = stash_controller_accounts
+        .into_iter()
+        .map(|(stash_account, controller_account)| {
+            let bond_call = compose_call!(
+                connection.metadata,
+                "Staking",
+                "bond",
+                GenericAddress::Id(controller_account.clone().to_owned()),
+                Compact(bond_value),
+                reward_destination.clone()
+            );
+            compose_call!(
+                connection.metadata,
+                "Sudo",
+                "sudo_as",
+                GenericAddress::Id(stash_account.clone().to_owned()),
+                bond_call
+            )
+        })
+        .collect::<Vec<_>>();
+
+    let xt = compose_extrinsic!(connection, "Utility", "batch", batch_bond_calls);
+    send_xt(
+        connection,
+        xt.hex_encode(),
+        "batch of bond calls",
+        XtStatus::InBlock,
+    );
+}
+
+pub fn nominate(connection: &Connection, nominee_key_pair: &KeyPair) {
+    let nominee_account_id = AccountId::from(nominee_key_pair.public());
+
+    let xt = connection.staking_nominate(vec![GenericAddress::Id(nominee_account_id)]);
+    send_xt(&connection, xt.hex_encode(), "nominate", XtStatus::InBlock);
+}
+
+pub fn batch_nominate(
+    connection: &Connection,
+    nominator_nominee_pairs: &[(&AccountId, &AccountId)],
+) {
+    let batch_nominate_calls = nominator_nominee_pairs
+        .iter()
+        .map(|(nominator, nominee)| {
+            let nominate_call = compose_call!(
+                connection.metadata,
+                "Staking",
+                "nominate",
+                vec![GenericAddress::Id(nominee.clone().to_owned())]
+            );
+            compose_call!(
+                connection.metadata,
+                "Sudo",
+                "sudo_as",
+                GenericAddress::Id(nominator.clone().to_owned()),
+                nominate_call
+            )
+        })
+        .collect::<Vec<_>>();
+
+    let xt = compose_extrinsic!(connection, "Utility", "batch", batch_nominate_calls);
+    send_xt(
+        connection,
+        xt.hex_encode(),
+        "batch of nominate calls",
+        XtStatus::InBlock,
+    );
+}
+
+pub fn bonded(connection: &Connection, stash: &KeyPair) -> Option<AccountId> {
+    let account_id = AccountId::from(stash.public());
+    connection
+        .get_storage_map("Staking", "Bonded", &account_id, None)
+        .expect(&format!(
+            "Failed to obtain Bonded for account id {}",
+            account_id
+        ))
+}
+
+pub fn ledger(
+    connection: &Connection,
+    controller: &KeyPair,
+) -> Option<pallet_staking::StakingLedger<AccountId, Balance>> {
+    let account_id = AccountId::from(controller.public());
+    connection
+        .get_storage_map("Staking", "Ledger", &account_id, None)
+        .expect(&format!(
+            "Failed to obtain Ledger for account id {}",
+            account_id
+        ))
 }
