+++ e2e-tests/src/lib.rs
@@ -1,6 +1,5 @@
 // TODO remove `pub`
 pub mod accounts;
 pub mod config;
-pub mod staking;
 pub mod test;
 pub mod transfer;
